Use ptrdiff_t for native-host offsets.
git-svn-id: svn://svn.code.sf.net/p/boomerang/code/trunk@2514 dc5e1367-85d4-4bd1-96b4-a20bfbfd92d5

+++ boomerang/db/prog.cpp
@@ -1339,7 +1339,7 @@ UserProc* Prog::getNextUserProc(std::list<Proc*>::iterator& it)
  * RETURNS:		Host pointer if in range; NULL if not
  *				Also sets 2 reference parameters (see above)
  *============================================================================*/
-const void* Prog::getCodeInfo(ADDRESS uAddr, const char*& last, int& delta)
+const void* Prog::getCodeInfo(ADDRESS uAddr, const char*& last, ptrdiff_t& delta)
 {
     delta=0;
     last=0;
