+++ boomerang/frontend/machine/hppa/simple.cpp
@@ -297,7 +297,7 @@ MATCH_finished_b:
 #line 91 "machine/hppa/simple.m"
 }
 
-int NJMCDecoder::decodeAssemblyInstruction (ADDRESS pc, int delta)
+int NJMCDecoder::decodeAssemblyInstruction (ADDRESS pc, ptrdiff_t delta)
 {
     char sCmplt[32];
     unsigned long r1,r2;
