+++ boomerang/frontend/machine/mc68k/decoder.cpp
@@ -62,7 +62,7 @@
  * RETURNS:        a DecodeResult structure containing all the information
  *                   gathered during decoding
  *============================================================================*/
-DecodeResult& NJMCDecoder::decodeInstruction (ADDRESS pc, int delta,
+DecodeResult& NJMCDecoder::decodeInstruction (ADDRESS pc, ptrdiff_t delta,
         UserProc* proc = NULL)
 {
     static DecodeResult result;
@@ -2737,7 +2737,7 @@ SemStr* NJMCDecoder::pAIndex(int d8, int r, int iT, int iR, int iS, int size)
     return ret;
 }
 
-SemStr* NJMCDecoder::pPcDisp(ADDRESS label, int delta, int size)
+SemStr* NJMCDecoder::pPcDisp(ADDRESS label, ptrdiff_t delta, int size)
 {
     // Note: label is in the host address space, so need to subtract delta
     SemStr* ret = new SemStr(size);
@@ -3122,7 +3122,7 @@ MATCH_finished_kd:
 
 
 SemStr* NJMCDecoder::amEAX(ADDRESS eax, ADDRESS x, DecodeResult& result,
-                           ADDRESS pc, int delta, int size)
+                           ADDRESS pc, ptrdiff_t delta, int size)
 {
     SemStr* ret;
     int reg2, mode;
@@ -3696,7 +3696,7 @@ MATCH_finished_ad:
 
 
 SemStr* NJMCDecoder::awlEAX(ADDRESS eax, ADDRESS x, DecodeResult& result,
-                            ADDRESS pc, int delta, int size)
+                            ADDRESS pc, ptrdiff_t delta, int size)
 {
     SemStr* ret;
     int reg2, mode;
@@ -4186,7 +4186,7 @@ MATCH_finished_rc:
 
 
 SemStr* NJMCDecoder::cEAX(ADDRESS eax, ADDRESS x, DecodeResult& result,
-                          ADDRESS pc, int delta, int size)
+                          ADDRESS pc, ptrdiff_t delta, int size)
 {
     SemStr* ret;
     int reg2, mode;
@@ -4576,7 +4576,7 @@ MATCH_finished_kc:
 
 
 SemStr* NJMCDecoder::dEAX(ADDRESS eax, ADDRESS x, DecodeResult& result,
-                          ADDRESS pc, int delta, int size)
+                          ADDRESS pc, ptrdiff_t delta, int size)
 {
     SemStr* ret;
     int reg2, mode;
@@ -5438,7 +5438,7 @@ MATCH_finished_vb:
 
 
 SemStr* NJMCDecoder::dBEAX(ADDRESS eax, ADDRESS x, DecodeResult& result,
-                           ADDRESS pc, int delta, int size)
+                           ADDRESS pc, ptrdiff_t delta, int size)
 {
     SemStr* ret;
     int reg2, mode;
@@ -5901,7 +5901,7 @@ MATCH_finished_nb:
 
 
 SemStr* NJMCDecoder::dWEAX(ADDRESS eax, ADDRESS x, DecodeResult& result,
-                           ADDRESS pc, int delta, int size)
+                           ADDRESS pc, ptrdiff_t delta, int size)
 {
     SemStr* ret;
     int reg2, mode;
@@ -6630,7 +6630,7 @@ MATCH_finished_ab:
 
 
 SemStr* NJMCDecoder::msEAX(ADDRESS eax, ADDRESS x, DecodeResult& result,
-                           ADDRESS pc, int delta, int size)
+                           ADDRESS pc, ptrdiff_t delta, int size)
 {
     SemStr* ret;
     int reg2, mode;
@@ -7511,7 +7511,7 @@ MATCH_finished_m:
 
 
 SemStr* NJMCDecoder::mrEAX(ADDRESS eax, ADDRESS x, DecodeResult& result,
-                           ADDRESS pc, int delta, int size)
+                           ADDRESS pc, ptrdiff_t delta, int size)
 {
     SemStr* ret;
     int reg2, mode;
