+++ boomerang/frontend/machine/sparc/disassembler.cpp
@@ -201,7 +201,7 @@ MATCH_finished_b:
  *					(i.e. number of bytes processed)
  *============================================================================*/
 
-int NJMCDecoder::decodeAssemblyInstruction (ADDRESS pc, int delta)
+int NJMCDecoder::decodeAssemblyInstruction (ADDRESS pc, ptrdiff_t delta)
 {
     ADDRESS hostPC = pc + delta;
     ADDRESS nextPC;
