+++ boomerang/frontend/machine/st20/decoder.m
@@ -51,7 +51,7 @@ void ST20Decoder::unused(int x)
  * RETURNS:		   a DecodeResult structure containing all the information gathered during decoding
  *============================================================================*/
 static	DecodeResult result;
-DecodeResult& ST20Decoder::decodeInstruction (ADDRESS pc, int delta) {
+DecodeResult& ST20Decoder::decodeInstruction (ADDRESS pc, ptrdiff_t delta) {
 	result.reset();							// Clear the result structure (numBytes = 0 etc)
 	ADDRESS hostPC = pc + delta;
 	std::list<Statement*>* stmts = NULL; 	// The actual list of instantiated Statements
@@ -347,6 +347,6 @@ ST20Decoder::ST20Decoder() : NJMCDecoder(prog)
 }
 
 // For now...
-int ST20Decoder::decodeAssemblyInstruction(ADDRESS, int)
+int ST20Decoder::decodeAssemblyInstruction(ADDRESS, ptrdiff_t)
 { return 0; }
