+++ boomerang/frontend/njmcDecoder.cpp
@@ -233,7 +233,7 @@ Exp* NJMCDecoder::dis_Num(unsigned num)
  * PARAMETERS:
  * RETURNS:			<none>
  *============================================================================*/
-void NJMCDecoder::unconditionalJump(const char* name, int size, ADDRESS relocd, int delta, ADDRESS pc,
+void NJMCDecoder::unconditionalJump(const char* name, int size, ADDRESS relocd, ptrdiff_t delta, ADDRESS pc,
                                     std::list<Statement*>* stmts, DecodeResult& result)
 {
     result.rtl = new RTL(pc, stmts);
