+++ boomerang/frontend/st20decoder.cpp
@@ -55,7 +55,7 @@ void ST20Decoder::unused(int x)
  * RETURNS:		   a DecodeResult structure containing all the information gathered during decoding
  *============================================================================*/
 static	DecodeResult result;
-DecodeResult& ST20Decoder::decodeInstruction (ADDRESS pc, int delta)
+DecodeResult& ST20Decoder::decodeInstruction (ADDRESS pc, ptrdiff_t delta)
 {
     result.reset();							// Clear the result structure (numBytes = 0 etc)
     ADDRESS hostPC = pc + delta;
@@ -1022,7 +1022,7 @@ ST20Decoder::ST20Decoder() : NJMCDecoder(prog)
 }
 
 // For now...
-int ST20Decoder::decodeAssemblyInstruction(ADDRESS, int)
+int ST20Decoder::decodeAssemblyInstruction(ADDRESS, ptrdiff_t)
 {
     return 0;
 }
