+++ boomerang/include/prog.h
@@ -134,7 +134,7 @@ public:
 
     // Lookup the given native address in the code section, returning a host pointer corresponding to the same
     // address
-    const void* getCodeInfo(ADDRESS uAddr, const char*& last, int& delta);
+    const void* getCodeInfo(ADDRESS uAddr, const char*& last, ptrdiff_t& delta);
 
     const char *getRegName(int idx)
     {
@@ -323,7 +323,7 @@ public:
         return pBF->readNative8(a);
     }
     Exp	  		*readNativeAs(ADDRESS uaddr, Type *type);
-    int			getTextDelta()
+    ptrdiff_t	getTextDelta()
     {
         return pBF->getTextDelta();
     }
