+++ boomerang/loader/BinaryFile.cpp
@@ -210,7 +210,7 @@ void BinaryFile::getTextLimits()
                         textDelta = (uintptr_t)pSect->uHostAddr - pSect->uNativeAddr;
                     else
                         {
-                            if (textDelta != (int) (pSect->uHostAddr - pSect->uNativeAddr))
+                            if (textDelta != (uintptr_t)pSect->uHostAddr - pSect->uNativeAddr)
                                 std::cerr << "warning: textDelta different for section " << pSect->pSectionName <<
                                           " (ignoring).\n";
                         }
